Refactored user to implement UserDetails directly. Security context holder principal is now the User object itself.
