+++ src/main/java/com/cognizant/training/model/User.java
@@ -2,9 +2,12 @@ package com.cognizant.training.model;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -18,7 +21,7 @@ import java.util.stream.Stream;
  */
 @Entity
 @Table(name = "users")
-public class User {
+public class User implements UserDetails {
 
     /**
      * The user's unique identifier.
@@ -108,12 +111,74 @@ public class User {
     }
 
     /**
-     * Gets all permissions granted discretely and by roles granted to the user.
+     * Gets all the authorities granted to the user. Combines authorities granted from roles and authorities granted
+     * directly to the user.
      *
-     * @return all permissions granted discretely and by roles granted to the user.
+     * @return all authorities granted to the user
      */
-    public Set<Permission> getAllPermissions() {
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
         return Stream.concat(permissions.stream(), getPermissionsFromRoles().stream())
                 .collect(Collectors.toSet());
+    }
+
+    /**
+     * Gets the password hash of the user
+     *
+     * @return password hash of the user
+     */
+    @Override
+    public String getPassword() {
+        return password;
+    }
+
+    /**
+     * Gets the username, also known as the email, of the user
+     *
+     * @return email of the user
+     */
+    @Override
+    public String getUsername() {
+        return email;
+    }
+
+    /**
+     * Whether the account is expired. Currently, accounts never expire.
+     *
+     * @return whether the account is non-expired.
+     */
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    /**
+     * Whether the account is locked. Currently, accounts can never lock.
+     *
+     * @return whether the account is non-locked
+     */
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    /**
+     * Whether the credentials are expired. Currently, account credentials never expire.
+     *
+     * @return whether the account credentials are non-expired
+     */
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    /**
+     * Whether the account is enabled. Currently, accounts are always enabled.
+     *
+     * @return whether the account is enabled.
+     */
+    @Override
+    public boolean isEnabled() {
+        return true;
     }
 }
