+++ src/main/java/com/cognizant/training/security/AuthenticationManagerImpl.java
@@ -1,5 +1,6 @@
 package com.cognizant.training.security;
 
+import com.cognizant.training.model.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -7,7 +8,6 @@ import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
@@ -25,14 +25,14 @@ public class AuthenticationManagerImpl implements AuthenticationManager {
 
     @Override
     public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        final UserDetails userDetails = userDetailsService.loadUserByUsername(authentication.getName());
-        if (!passwordEncoder().matches(authentication.getCredentials().toString(), userDetails.getPassword())) {
+        final User user = (User) userDetailsService.loadUserByUsername(authentication.getName());
+        if (!passwordEncoder().matches(authentication.getCredentials().toString(), user.getPassword())) {
             throw new BadCredentialsException("invalid username or password");
         }
 
         return new UsernamePasswordAuthenticationToken(
-                userDetails.getUsername(),
-                userDetails.getPassword(),
-                userDetails.getAuthorities());
+                user,
+                user.getPassword(),
+                user.getAuthorities());
     }
 }
