+++ src/main/java/com/cognizant/training/service/UserService.java
@@ -3,10 +3,8 @@ package com.cognizant.training.service;
 import com.cognizant.training.exception.LoginException;
 import com.cognizant.training.exception.RegisterException;
 import com.cognizant.training.request.LoginRequest;
-import com.cognizant.training.security.UserDetailsImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -26,6 +24,7 @@ public class UserService implements IUserService {
         Authentication authentication = authenticationManager.authenticate(
                 new UsernamePasswordAuthenticationToken(request.getEmail(), request.getPassword())
         );
+
         SecurityContextHolder.getContext().setAuthentication(authentication);
     }
