+++ src/main/java/com/cognizant/training/util/JwtUtils.java
@@ -13,6 +13,8 @@ import javax.annotation.PostConstruct;
 import java.security.Key;
 import java.util.Base64;
 import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 
 @Component
 public class JwtUtils {
@@ -80,8 +82,11 @@ public class JwtUtils {
     }
 
     public String generateJwtToken(User user) {
+        Map<String, Object> claims = new HashMap<>();
+        claims.put("email", user.getEmail());
         return Jwts.builder()
                 .setSubject(user.getId().toString())
+                .setClaims(claims)
                 .setIssuedAt(new Date())
                 .setExpiration(new Date((new Date()).getTime() + expiration))
                 .signWith(key)
